This commit corrects two errors in the base code provided by the course instructor.
The first problem was that the provided method "first_nonzero_index" expects to receive an iterable as input yet there were a couple places in the code where a Vector type was passed to it.  This resulted in "TypeError: 'Vector' object is not iterable" error messages.  I fixed that by passing the .coordinates attribute instead of the entire Vector.

The second problem was in the set_basepoint method.  Input from the original_index value (which could've been any numeric type) was being divided by decimal.Decimal type.  This resulted in the "TypeError: unsupported operand type(s) for /: 'decimal.Decimal' and 'float'" error.  I fixed that by casting the offending input as a Decimal type.

With these two corrections I am now able to create Line instances without raising any errors.  The code is now suitable for me to add my methods to it.